bar and dot plot of effect size of microbiomeMarker data

bar and dot plot of effect size microbiomeMarker data. This function returns a ggplot2 object that can be saved or further customized using ggplot2 package.

plot_ef_bar(mm, label_level = 1, max_label_len = 60, markers = NULL)

plot_ef_dot(mm, label_level = 1, max_label_len = 60, markers = NULL)

Arguments

mm
a microbiomeMarker object

label_level
integer, number of label levels to be displayed, default 1, 0 means display the full name of the feature

max_label_len
integer, maximum number of characters of feature label, default 60

markers
character vector, markers to display, default NULL, indicating plot all markers.

Value

a ggplot project