Import function to read the the output of dada2 as phyloseq object

Import the output of dada2 into phyloseq object

import_dada2(
  seq_tab,
  tax_tab = NULL,
  sam_tab = NULL,
  phy_tree = NULL,
  keep_taxa_rows = TRUE
)

Arguments

seq_tab
matrix-like, ASV table, the output of dada2::removeBimeraDenovo.

tax_tab
matrix, taxonomy table, the output of dada2::assignTaxonomy or dada2::addSpecies.

sam_tab
data.frame or phyloseq::sample_data, sample data

phy_tree
ape::phylo class or character represents the path of the tree file

keep_taxa_rows
logical, whether keep taxa in rows or not in the otu_table of the returned phyloseq object, default TRUE.

Value

phyloseq::phyloseq object hold the taxonomy info, sample metadata, number of reads per ASV.

Details

The output of the dada2 pipeline is a feature table of amplicon sequence variants (an ASV table): A matrix with rows corresponding to samples and columns to ASVs, in which the value of each entry is the number of times that ASV was observed in that sample. This table is analogous to the traditional OTU table. Conveniently, taxa names are saved as ASV1, ASV2, ..., in the returned phyloseq object.

Examples

seq_tab <- readRDS(system.file("extdata", "dada2_seqtab.rds",
    package = "microbiomeMarker"
))
tax_tab <- readRDS(system.file("extdata", "dada2_taxtab.rds",
    package = "microbiomeMarker"
))
sam_tab <- read.table(system.file("extdata", "dada2_samdata.txt",
    package = "microbiomeMarker"
), sep = "\t", header = TRUE, row.names = 1)
ps <- import_dada2(seq_tab = seq_tab, tax_tab = tax_tab, sam_tab = sam_tab)
ps
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 232 taxa and 20 samples ]
#> sample_data() Sample Data:       [ 20 samples by 4 sample variables ]
#> tax_table()   Taxonomy Table:    [ 232 taxa by 6 taxonomic ranks ]
#> refseq()      DNAStringSet:      [ 232 reference sequences ]