Convert phyloseq data to MetagenomeSeq MRexperiment object

The phyloseq data is converted to the relevant metagenomeSeq::MRexperiment object, which can then be tested in the zero-inflated mixture model framework in the metagenomeSeq package.

phyloseq2metagenomeSeq(ps, ...)

otu_table2metagenomeSeq(ps, ...)

Arguments

ps

phyloseq::phyloseq object for phyloseq2metagenomeSeq(), or phyloseq::otu_table object for otu_table2metagenomeseq().

...

optional, additional named arguments passed to metagenomeSeq::newMRexperiment(). Most users will not need to pass any additional arguments here.

Value

A metagenomeSeq::MRexperiment object.

See also

Examples

data(caporaso)
phyloseq2metagenomeSeq(caporaso)
#> MRexperiment (storageMode: environment)
#> assayData: 3426 features, 34 samples 
#>   element names: counts 
#> protocolData: none
#> phenoData
#>   sampleNames: L1S140 L1S208 ... L2S240 (34 total)
#>   varLabels: SampleType Year ... Description (8 total)
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: New.CleanUp.ReferenceOTU647 14030 ... 3769033 (3426
#>     total)
#>   fvarLabels: OTUname Kingdom ... Species (8 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation: