Subset microbiome markers

Subset markers based on an expression related to the columns and values within the marker_table slot of mm.

subset_marker(mm, ...)

Arguments

mm

a microbiomeMarker or marker_table object.

...

the subsetting expression passed to base::subset().

Value

a subset object in the same class with mm.

Examples

data(enterotypes_arumugam)
mm <- run_limma_voom(
    enterotypes_arumugam,
    "Enterotype",
    contrast = c("Enterotype 3", "Enterotype 2"),
    pvalue_cutoff = 0.01,
    p_adjust = "none"
)
subset_marker(mm, pvalue < 0.005)
#> microbiomeMarker-class inherited from phyloseq-class
#> normalization method:              [ none ]
#> microbiome marker identity method: [ limma_voom ]
#> marker_table() Marker Table:       [ 8 microbiome markers with 5 variables ]
#> otu_table()    OTU Table:          [ 244 taxa and  39 samples ]
#> tax_table()    Taxonomy Table:     [ 244 taxa by 1 taxonomic ranks ]