Summary differential analysis methods comparison results

Summary differential analysis methods comparison results

# S3 method for compareDA
summary(
  object,
  sort = c("score", "auc", "fpr", "power"),
  boot = TRUE,
  boot_n = 1000L,
  prob = c(0.05, 0.95),
  ...
)

Arguments

object

an compareDA object, output from compare_DA().

sort

character string specifying sort method. Possibilities are "score" which is calculated as \((auc - 0.5) * power - fdr\), "auc" for area under the ROC curve, "fpr" for false positive rate, "power" for empirical power.

boot

logical, whether use bootstrap for confidence limites of the score, default TRUE. Recommended to be TRUE unless n_rep is larger then 100 in compare_DA().

boot_n

integer, number of bootstraps, default 1000L.

prob

two length numeric vector, confidence limits for score, default c(0.05, 0.95).

...

extra arguments affecting the summary produced.

Value

a data.frame containing measurements for differential analysis methods:

call: differential analysis commands.

auc: area under curve of ROC.

fpr: false positive rate

power: empirical power.

fdr: false discover7y rate.

score: score whch is calculated as \((auc - 0.5) * power - fdr\).

score_*: confidence limits of score.